+++ HPLLib/HPLStockManager.cpp
@@ -322,7 +322,7 @@ bool hpl::aleph::HPLStockManager::deleteSide(int index)
 }
 bool hpl::aleph::HPLStockManager::deleteObject(int index)
 {
-    assert(index >= 0 && index < SavedObjectList.size());
+    assert(index >= 0 && index < (int)SavedObjectList.size());
     this->delObjects[index] = true;
     updateDeletes();
     return true;
